Moved ArchiveWSServerHelper to shared subdirectory
