+++ dgva-server/src/main/java/uk/ac/ebi/dgva/server/ws/ArchiveWSServer.java
@@ -28,8 +28,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import uk.ac.ebi.eva.lib.metadata.dgva.ArchiveDgvaDBAdaptor;
 import uk.ac.ebi.eva.lib.metadata.dgva.StudyDgvaDBAdaptor;
-import uk.ac.ebi.eva.lib.metadata.ArchiveWSServerHelper;
-import uk.ac.ebi.eva.lib.utils.QueryOptions;
+import uk.ac.ebi.eva.lib.metadata.shared.ArchiveWSServerHelper;
 import uk.ac.ebi.eva.lib.utils.QueryResponse;
 
 import java.io.IOException;
@@ -51,18 +50,18 @@ public class ArchiveWSServer extends DgvaWSServer {
 
 
     public ArchiveWSServer() throws IOException {
-        archiveWSServerHelper = new ArchiveWSServerHelper(this.version);
+        archiveWSServerHelper = new ArchiveWSServerHelper();
     }
 
     @RequestMapping(value = "/studies/all", method = RequestMethod.GET)
     public QueryResponse getStudies(@RequestParam(name = "species", required = false) List<String> species,
                                     @RequestParam(name = "type", required = false) List<String> types) {
-        return archiveWSServerHelper.getStudies(species, types, queryUtils, studyDgvaDbAdaptor);
+        return archiveWSServerHelper.getStudies(species, types, queryUtils, studyDgvaDbAdaptor, version);
     }
 
     @RequestMapping(value = "/studies/stats", method = RequestMethod.GET)
     public QueryResponse getStudiesStats(@RequestParam(name = "species", required = false) List<String> species,
                                          @RequestParam(name = "type", required = false) List<String> types) {
-        return archiveWSServerHelper.getStudiesStats(species, types, queryUtils, archiveDgvaDbAdaptor);
+        return archiveWSServerHelper.getStudiesStats(species, types, queryUtils, archiveDgvaDbAdaptor, version);
     }
 }
