+++ eva-server/src/main/java/uk/ac/ebi/eva/server/ws/ArchiveWSServer.java
@@ -21,24 +21,24 @@ package uk.ac.ebi.eva.server.ws;
 
 import io.swagger.annotations.Api;
 
-import uk.ac.ebi.eva.commons.mongodb.entities.projections.VariantStudySummary;
-import uk.ac.ebi.eva.lib.utils.QueryResponse;
-import uk.ac.ebi.eva.lib.utils.QueryResult;
-import uk.ac.ebi.eva.lib.utils.QueryOptions;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+
+import uk.ac.ebi.eva.commons.mongodb.entities.projections.VariantStudySummary;
 import uk.ac.ebi.eva.commons.mongodb.services.VariantStudySummaryService;
+
 import uk.ac.ebi.eva.lib.metadata.eva.ArchiveEvaproDBAdaptor;
-import uk.ac.ebi.eva.lib.metadata.ArchiveWSServerHelper;
 import uk.ac.ebi.eva.lib.metadata.eva.StudyEvaproDBAdaptor;
+import uk.ac.ebi.eva.lib.metadata.shared.ArchiveWSServerHelper;
 import uk.ac.ebi.eva.lib.eva_utils.DBAdaptorConnector;
 import uk.ac.ebi.eva.lib.eva_utils.MultiMongoDbFactory;
+import uk.ac.ebi.eva.lib.utils.QueryResponse;
+import uk.ac.ebi.eva.lib.utils.QueryResult;
 
 import java.io.IOException;
-import java.util.Arrays;
 import java.util.List;
 
 @RestController
@@ -58,7 +58,7 @@ public class ArchiveWSServer extends EvaWSServer {
     private ArchiveWSServerHelper archiveWSServerHelper;
 
     public ArchiveWSServer() {
-        archiveWSServerHelper = new ArchiveWSServerHelper(this.version);
+        archiveWSServerHelper = new ArchiveWSServerHelper();
     }
 
     @RequestMapping(value = "/files/count", method = RequestMethod.GET)
@@ -84,7 +84,7 @@ public class ArchiveWSServer extends EvaWSServer {
     @RequestMapping(value = "/studies/all", method = RequestMethod.GET)
     public QueryResponse getStudies(@RequestParam(name = "species", required = false) List<String> species,
                                     @RequestParam(name = "type", required = false) List<String> types) {
-        return archiveWSServerHelper.getStudies(species, types, queryUtils, studyEvaproDbAdaptor);
+        return archiveWSServerHelper.getStudies(species, types, queryUtils, studyEvaproDbAdaptor, version);
     }
 
     @RequestMapping(value = "/studies/list", method = RequestMethod.GET)
@@ -99,6 +99,6 @@ public class ArchiveWSServer extends EvaWSServer {
     @RequestMapping(value = "/studies/stats", method = RequestMethod.GET)
     public QueryResponse getStudiesStats(@RequestParam(name = "species", required = false) List<String> species,
                                          @RequestParam(name = "type", required = false) List<String> types) {
-        return archiveWSServerHelper.getStudiesStats(species, types, queryUtils, archiveEvaproDbAdaptor);
+        return archiveWSServerHelper.getStudiesStats(species, types, queryUtils, archiveEvaproDbAdaptor, version);
     }
 }
